Merge "Remove module-metadata-include-filter in mts-dnsresolver.xml" am: d7c074b109 am: c515c44c85 am: 869531e752 am: e73da0b2a4 am: a799a4567a
This is to solve the issue that the # of tests executed for
dns resolver is not correct.

module-metadata-include-filter is used to filter out test
that doesn't match the component in AndroidTest.xml. Hence,
the resolv_* are filtered out because its component doesn't
match "networking"

Bug: 150834981

Test: mts-tradefed run mts -m mts-dnsresolver

Change-Id: Ibb543577d87b11e2e04f5b26b5723202d877a3be
(cherry picked from commit a4428e7fb98ce989b07de34cdbeaeb2d11153a98)