Create mts-angle-gce.xml
We need to run CTS tests on Cuttlefish to collect code coverage numbers,
however we need to run additional tests versus real devices to get the
native code coverage.   We aren't ready to enable these tests on real
devices yet due to bugs in the underlying Vulkan drivers that ANGLE
uses.

The contents of this are a translation of the following .gcl files:
https://source.corp.google.com/piper///depot/google3/wireless/android/test_tools/atp/config/test/cts/graphics/angle-host-test.gcl
https://source.corp.google.com/piper///depot/google3/wireless/android/test_tools/atp/config/test/cts/graphics/angle-deqp.gcl

Bug: 150464273
Test: None (no way to test this until it's checked in)
Change-Id: Ic5fd6cb40312c9a9bacc936e67309460499c9c82