Use MTS configuration for network MTS tests
MTS is currently built with AOSP certificates, but run on devices that
use different certificates. Use a special target that overrides the
certificate as a workaround, considering that there is no short-term
solution to this problem.

This configuration will not work when run on AOSP devices, so it should
be reverted when pushed to AOSP, until there is a way to use the same
target in all branches.

Bug: 154366165
Test: mts-networking-coverage remote run
Change-Id: Ibe6f881fc235e7d78da6870ebaa329a4ba5950e5