Add ART gtests to the ART Module test plan for MTS.
Also, following the MTS documentation, split the list of ART tests in
two sub-lists:
- a list of tests that do not need root
  access (`tools/mts-tradefed/res/config/mts-art-tests-list-user.xml`);
- a list of tests that do need root
  access (`tools/mts-tradefed/res/config/mts-art-tests-list-eng-only.xml`).

The currently supported ART run-tests do not need root access and thus
are all part of the first ("user") list. As many ART gtests require
root access, all ART gtests have been included in the
second ("eng-only") list for the sake of simplicity, but we should
later identify the ones that do not actually require root access and
move them from the "eng-only" list to the "user" list.

Test: m mts && mts-tradefed run commandAndExit mts-art
Bug: 167385698
Change-Id: Ic47be897b26cae722a89c812239f977a8958587e