Add CtsStrictJavaPackagesTestCases to MTS.
The test has been tagged as "mts" before but has not been added to any
config.

Note that the test already runs as a global presubmit test on multiple
branches. However, we would want the test to be run by partners as
well, as it guarantees correctness of BOOTCLASSPATH and
SYSTEMSERVERCLASSPATH env variables.

Bug: 180105615
Test: m mts && mts-tradefed run mts-core
Change-Id: Ie0fe7d8df2b7f1c549f607381b1b8a7dad3a911c