Remove the notification Cts from mts-smoke-tests-list.xml
In order to avoid updating mainline modules to break APIs, it needs
to add those cts included the logic in ExtServices module. After
checking, the notification cts don't include the logic in ExtServices
module, it should be ok to remove them from MTS because we don't
violate the mainline goal.

Bug: 182959140
Bug: 181388339
Test: Build
Change-Id: I78b911167202b59646d072b1b3023c32ef24e6e2