Regenerate the ART MTS definition to exclude test `NativeCryptoArgTest`.
Test `com.android.org.conscrypt.NativeCryptoArgTest` requires an
updated Conscrypt Module (which may not be available on some test
configurations) to run properly and is not really relevant to ART
Module testing. Exclude it from ART MTS for now, as this is deemed
simpler than the alternative (installing an up-to-date Conscrypt
Module as a prerequisite to running ART Module tests).

Test: m mts && mts-tradefed run commandAndExit mts-art
Test: m mts && mts-tradefed run commandAndExit mts-art-shard-01
Bug: b/198606382
Change-Id: Ifbb088244346ae0bb36f8188795496c3119b39dd