Add CtsHostideNetworkTests to MTS.
This test is already in CTS, and it should run in MTS as well
because it tests things like VPN APIs that could be affected by
changes in the Tethering module.

Previously this module contained lots of tests for platform code,
but those were recently moved back to the cts/ project.

Also reorder the stanzas to first list the filters, then the
excludes, then the module args.

Bug: 358007993
Test: test-only change
Change-Id: I372601e3ee0dec552cdfb66771b5c1d12d497ba3